reafactor: rewrite cjsToEsm.cjs to esm.mjs
